Move SAA scheduling to appointments controller
from the over-used `ServiceProviderReferralsController`.

This also refactors the functions to make them longer but less DRY, as
the re-use of various private functions made it tricky to understand
what was going on at a glance and to update.
