+++ server/routes/appointments/appointmentsController.ts
@@ -1,6 +1,8 @@
 import { Request, Response } from 'express'
+import SupplierAssessmentDecorator from '../../decorators/supplierAssessmentDecorator'
 import { AppointmentSchedulingDetails } from '../../models/appointment'
 import DeliusOfficeLocation from '../../models/deliusOfficeLocation'
+import AuthUserDetails from '../../models/hmppsAuth/authUserDetails'
 import CommunityApiService from '../../services/communityApiService'
 import DeliusOfficeLocationFilter from '../../services/deliusOfficeLocationFilter'
 import DraftsService from '../../services/draftsService'
@@ -11,9 +13,15 @@ import ControllerUtils from '../../utils/controllerUtils'
 import { FormValidationError } from '../../utils/formValidationError'
 import ScheduleActionPlanSessionPresenter from '../service-provider/action-plan/sessions/edit/scheduleActionPlanSessionPresenter'
 import ActionPlanSessionCheckAnswersPresenter from '../serviceProviderReferrals/actionPlanSessionCheckAnswersPresenter'
+import InitialAssessmentCheckAnswersPresenter from '../serviceProviderReferrals/initialAssessmentCheckAnswersPresenter'
 import ScheduleAppointmentCheckAnswersView from '../serviceProviderReferrals/scheduleAppointmentCheckAnswersView'
 import ScheduleAppointmentForm from '../serviceProviderReferrals/scheduleAppointmentForm'
+import ScheduleAppointmentPresenter from '../serviceProviderReferrals/scheduleAppointmentPresenter'
 import ScheduleAppointmentView from '../serviceProviderReferrals/scheduleAppointmentView'
+import SupplierAssessmentAppointmentConfirmationPresenter from '../serviceProviderReferrals/supplierAssessmentAppointmentConfirmationPresenter'
+import SupplierAssessmentAppointmentConfirmationView from '../serviceProviderReferrals/supplierAssessmentAppointmentConfirmationView'
+import SupplierAssessmentAppointmentPresenter from '../shared/supplierAssessmentAppointmentPresenter'
+import SupplierAssessmentAppointmentView from '../shared/supplierAssessmentAppointmentView'
 import AppointmentSummary from './appointmentSummary'
 
 export type DraftAppointmentBooking = null | AppointmentSchedulingDetails
@@ -29,6 +37,203 @@ export default class AppointmentsController {
     private readonly referenceDataService: ReferenceDataService
   ) {
     this.deliusOfficeLocationFilter = new DeliusOfficeLocationFilter(referenceDataService)
+  }
+
+  async startSupplierAssessmentAppointmentScheduling(req: Request, res: Response): Promise<void> {
+    const draftBooking = await this.draftsService.createDraft('supplierAssessmentBooking', null, {
+      userId: res.locals.user.userId,
+    })
+
+    res.redirect(`/service-provider/referrals/${req.params.id}/supplier-assessment/schedule/${draftBooking.id}/details`)
+  }
+
+  async scheduleSupplierAssessmentAppointment(req: Request, res: Response): Promise<void> {
+    const fetchResult = await this.fetchDraftBookingOrRenderMessage(req, res)
+    if (fetchResult.rendered) {
+      return
+    }
+    const { draft } = fetchResult
+
+    const referralId = req.params.id
+    const { accessToken } = res.locals.user.token
+    const referral = await this.interventionsService.getSentReferral(accessToken, referralId)
+    const supplierAssessment = await this.interventionsService.getSupplierAssessment(accessToken, referralId)
+    const intervention = await this.interventionsService.getIntervention(accessToken, referral.referral.interventionId)
+    const deliusOfficeLocations: DeliusOfficeLocation[] =
+      await this.deliusOfficeLocationFilter.findOfficesByIntervention(intervention)
+    const { currentAppointment } = new SupplierAssessmentDecorator(supplierAssessment)
+    const hasExistingScheduledAppointment = currentAppointment !== null && !currentAppointment.sessionFeedback.submitted
+
+    let userInputData: Record<string, unknown> | null = null
+    let formError: FormValidationError | null = null
+    let serverError: FormValidationError | null = null
+
+    if (req.method === 'POST') {
+      const data = await new ScheduleAppointmentForm(req, deliusOfficeLocations).data()
+      if (data.error) {
+        res.status(400)
+        formError = data.error
+        userInputData = req.body
+      } else {
+        await this.draftsService.updateDraft(draft.id, data.paramsForUpdate, { userId: res.locals.user.userId })
+        res.redirect(`/service-provider/referrals/${referralId}/supplier-assessment/schedule/${draft.id}/check-answers`)
+        return
+      }
+    } else if (req.query.clash === 'true') {
+      serverError = {
+        errors: [
+          {
+            formFields: ['session-input'],
+            errorSummaryLinkedField: 'session-input',
+            message:
+              'The proposed date and time you selected clashes with another appointment. Please select a different date and time.',
+          },
+        ],
+      }
+    }
+
+    const serviceUser = await this.communityApiService.getServiceUserByCRN(referral.referral.serviceUser.crn)
+
+    let assignedCaseworker: AuthUserDetails | null = null
+    if (currentAppointment?.sessionFeedback?.submitted) {
+      assignedCaseworker = await this.hmppsAuthService.getSPUserByUsername(
+        res.locals.user.token.accessToken,
+        currentAppointment!.sessionFeedback!.submittedBy!.username
+      )
+    }
+    let appointmentSummary: AppointmentSummary | null = null
+    if (currentAppointment) {
+      const deliusOfficeLocation = await this.deliusOfficeLocationFilter.findOfficeByAppointment(currentAppointment)
+      appointmentSummary = new AppointmentSummary(currentAppointment, assignedCaseworker, deliusOfficeLocation)
+    }
+    const overrideBackLinkHref = hasExistingScheduledAppointment
+      ? `/service-provider/referrals/${referralId}/supplier-assessment`
+      : undefined
+
+    const presenter = new ScheduleAppointmentPresenter(
+      'supplierAssessment',
+      referral,
+      currentAppointment,
+      appointmentSummary,
+      deliusOfficeLocations,
+      formError,
+      draft.data,
+      userInputData,
+      serverError,
+      overrideBackLinkHref
+    )
+
+    const view = new ScheduleAppointmentView(presenter)
+    ControllerUtils.renderWithLayout(res, view, serviceUser)
+  }
+
+  async checkSupplierAssessmentAnswers(req: Request, res: Response): Promise<void> {
+    const referralId = req.params.id
+
+    const referral = await this.interventionsService.getSentReferral(res.locals.user.token.accessToken, referralId)
+    const serviceUser = await this.communityApiService.getServiceUserByCRN(referral.referral.serviceUser.crn)
+
+    const fetchResult = await this.fetchDraftBookingOrRenderMessage(req, res)
+    if (fetchResult.rendered) {
+      return
+    }
+
+    const presenter = new InitialAssessmentCheckAnswersPresenter(fetchResult.draft, referral.id)
+    const view = new ScheduleAppointmentCheckAnswersView(presenter)
+
+    ControllerUtils.renderWithLayout(res, view, serviceUser)
+  }
+
+  async submitSupplierAssessmentAppointment(req: Request, res: Response): Promise<void> {
+    const referralId = req.params.id
+    const { draftBookingId } = req.params
+    const { accessToken } = res.locals.user.token
+    const supplierAssessment = await this.interventionsService.getSupplierAssessment(accessToken, referralId)
+
+    const { currentAppointment } = new SupplierAssessmentDecorator(supplierAssessment)
+    const hasExistingScheduledAppointment = currentAppointment !== null && !currentAppointment.sessionFeedback.submitted
+
+    const fetchResult = await this.fetchDraftBookingOrRenderMessage(req, res)
+    if (fetchResult.rendered) {
+      return
+    }
+    const { draft } = fetchResult
+
+    if (draft.data === null) {
+      throw new Error('Draft data was unexpectedly null when submitting appointment')
+    }
+
+    try {
+      await this.interventionsService.scheduleSupplierAssessmentAppointment(
+        res.locals.user.token.accessToken,
+        supplierAssessment.id,
+        draft.data
+      )
+    } catch (e) {
+      const interventionsServiceError = e as InterventionsServiceError
+      if (interventionsServiceError.status === 409) {
+        res.redirect(
+          `/service-provider/referrals/${req.params.id}/supplier-assessment/schedule/${draftBookingId}/details?clash=true`
+        )
+        return
+      }
+
+      throw e
+    }
+
+    await this.draftsService.deleteDraft(draft.id, { userId: res.locals.user.userId })
+
+    const successfulRedirectPath = hasExistingScheduledAppointment
+      ? 'rescheduled-confirmation'
+      : 'scheduled-confirmation'
+
+    res.redirect(`/service-provider/referrals/${referralId}/supplier-assessment/${successfulRedirectPath}`)
+  }
+
+  async showSupplierAssessmentAppointmentConfirmation(
+    req: Request,
+    res: Response,
+    { isReschedule }: { isReschedule: boolean }
+  ): Promise<void> {
+    const referralId = req.params.id
+
+    const referral = await this.interventionsService.getSentReferral(res.locals.user.token.accessToken, referralId)
+    const serviceUser = await this.communityApiService.getServiceUserByCRN(referral.referral.serviceUser.crn)
+
+    const presenter = new SupplierAssessmentAppointmentConfirmationPresenter(referral, isReschedule)
+    const view = new SupplierAssessmentAppointmentConfirmationView(presenter)
+
+    return ControllerUtils.renderWithLayout(res, view, serviceUser)
+  }
+
+  async showSupplierAssessmentAppointment(req: Request, res: Response): Promise<void> {
+    const referralId = req.params.id
+
+    const [referral, supplierAssessment] = await Promise.all([
+      this.interventionsService.getSentReferral(res.locals.user.token.accessToken, referralId),
+
+      this.interventionsService.getSupplierAssessment(res.locals.user.token.accessToken, referralId),
+    ])
+
+    const appointment = new SupplierAssessmentDecorator(supplierAssessment).currentAppointment
+    if (appointment === null) {
+      throw new Error('Attempting to view supplier assessment without a current appointment')
+    }
+    const deliusOfficeLocation = await this.deliusOfficeLocationFilter.findOfficeByAppointment(appointment)
+
+    const serviceUser = await this.communityApiService.getServiceUserByCRN(referral.referral.serviceUser.crn)
+
+    const presenter = new SupplierAssessmentAppointmentPresenter(
+      referral,
+      appointment,
+      new AppointmentSummary(appointment, null, deliusOfficeLocation),
+      {
+        userType: 'service-provider',
+      }
+    )
+    const view = new SupplierAssessmentAppointmentView(presenter)
+
+    return ControllerUtils.renderWithLayout(res, view, serviceUser)
   }
 
   async startEditingActionPlanSessionAppointment(req: Request, res: Response): Promise<void> {
