+++ server/routes/serviceProviderReferrals/serviceProviderReferralsController.ts
@@ -28,9 +28,6 @@ import ActionPlanConfirmationView from '../service-provider/action-plan/confirma
 import AddActionPlanNumberOfSessionsView from '../service-provider/action-plan/number-of-sessions/actionPlanNumberOfSessionsView'
 import AddActionPlanNumberOfSessionsPresenter from '../service-provider/action-plan/number-of-sessions/actionPlanNumberOfSessionsPresenter'
 import ActionPlanNumberOfSessionsForm from '../service-provider/action-plan/number-of-sessions/actionPlanNumberOfSessionsForm'
-import ScheduleAppointmentPresenter from './scheduleAppointmentPresenter'
-import ScheduleAppointmentView from './scheduleAppointmentView'
-import ScheduleAppointmentForm from './scheduleAppointmentForm'
 import AttendanceFeedbackView from '../appointments/feedback/shared/attendance/attendanceFeedbackView'
 import ActionPlanPostSessionAttendanceFeedbackPresenter from '../appointments/feedback/actionPlanSessions/attendance/actionPlanPostSessionAttendanceFeedbackPresenter'
 import AttendanceFeedbackForm from '../appointments/feedback/shared/attendance/attendanceFeedbackForm'
@@ -57,12 +54,7 @@ import ServiceCategory from '../../models/serviceCategory'
 import AssessRisksAndNeedsService from '../../services/assessRisksAndNeedsService'
 import ActionPlanPresenter from '../shared/action-plan/actionPlanPresenter'
 import ActionPlanView from '../shared/action-plan/actionPlanView'
-import SentReferral from '../../models/sentReferral'
 import SupplierAssessmentDecorator from '../../decorators/supplierAssessmentDecorator'
-import SupplierAssessmentAppointmentPresenter from '../shared/supplierAssessmentAppointmentPresenter'
-import SupplierAssessmentAppointmentView from '../shared/supplierAssessmentAppointmentView'
-import SupplierAssessmentAppointmentConfirmationPresenter from './supplierAssessmentAppointmentConfirmationPresenter'
-import SupplierAssessmentAppointmentConfirmationView from './supplierAssessmentAppointmentConfirmationView'
 import ActionPlanEditConfirmationPresenter from '../service-provider/action-plan/edit/actionPlanEditConfirmationPresenter'
 import ActionPlanEditConfirmationView from '../service-provider/action-plan/edit/actionPlanEditConfirmationView'
 import InitialAssessmentAttendanceFeedbackPresenter from '../appointments/feedback/initialAssessment/attendance/initialAssessmentAttendanceFeedbackPresenter'
@@ -71,19 +63,15 @@ import InitialAssessmentFeedbackConfirmationPresenter from '../appointments/feed
 import InitialAssessmentFeedbackConfirmationView from '../appointments/feedback/initialAssessment/confirmation/initialAssessmentFeedbackConfirmationView'
 import ActionPlanSessionBehaviourFeedbackPresenter from '../appointments/feedback/actionPlanSessions/behaviour/actionPlanSessionBehaviourFeedbackPresenter'
 import InitialAssessmentBehaviourFeedbackPresenter from '../appointments/feedback/initialAssessment/behaviour/initialAssessmentBehaviourFeedbackPresenter'
-import DraftsService, { Draft } from '../../services/draftsService'
-import AuthUserDetails from '../../models/hmppsAuth/authUserDetails'
+import DraftsService from '../../services/draftsService'
 import {
   ActionPlanAppointment,
   AppointmentSchedulingDetails,
   InitialAssessmentAppointment,
 } from '../../models/appointment'
-import DeliusOfficeLocation from '../../models/deliusOfficeLocation'
 import DeliusOfficeLocationFilter from '../../services/deliusOfficeLocationFilter'
 import AppointmentSummary from '../appointments/appointmentSummary'
 import ReferenceDataService from '../../services/referenceDataService'
-import InitialAssessmentCheckAnswersPresenter from './initialAssessmentCheckAnswersPresenter'
-import ScheduleAppointmentCheckAnswersView from './scheduleAppointmentCheckAnswersView'
 import createFormValidationErrorOrRethrow from '../../utils/interventionsFormError'
 import EndOfServiceReportPresenter from '../shared/endOfServiceReport/endOfServiceReportPresenter'
 import EndOfServiceReportView from '../shared/endOfServiceReport/endOfServiceReportView'
@@ -602,266 +590,6 @@ export default class ServiceProviderReferralsController {
     const view = new AddActionPlanNumberOfSessionsView(presenter)
     res.status(formError === null ? 200 : 400)
     return ControllerUtils.renderWithLayout(res, view, serviceUser)
-  }
-
-  async showSupplierAssessmentAppointmentConfirmation(
-    req: Request,
-    res: Response,
-    { isReschedule }: { isReschedule: boolean }
-  ): Promise<void> {
-    const referralId = req.params.id
-
-    const referral = await this.interventionsService.getSentReferral(res.locals.user.token.accessToken, referralId)
-    const serviceUser = await this.communityApiService.getServiceUserByCRN(referral.referral.serviceUser.crn)
-
-    const presenter = new SupplierAssessmentAppointmentConfirmationPresenter(referral, isReschedule)
-    const view = new SupplierAssessmentAppointmentConfirmationView(presenter)
-
-    return ControllerUtils.renderWithLayout(res, view, serviceUser)
-  }
-
-  async startSupplierAssessmentAppointmentScheduling(req: Request, res: Response): Promise<void> {
-    const draftBooking = await this.draftsService.createDraft('supplierAssessmentBooking', null, {
-      userId: res.locals.user.userId,
-    })
-
-    res.redirect(`/service-provider/referrals/${req.params.id}/supplier-assessment/schedule/${draftBooking.id}/details`)
-  }
-
-  async scheduleSupplierAssessmentAppointment(req: Request, res: Response): Promise<void> {
-    const fetchResult = await this.fetchDraftBookingOrRenderMessage(req, res)
-    if (fetchResult.rendered) {
-      return
-    }
-    const { draft } = fetchResult
-
-    const referralId = req.params.id
-    const { accessToken } = res.locals.user.token
-    const referral = await this.interventionsService.getSentReferral(accessToken, referralId)
-    const supplierAssessment = await this.interventionsService.getSupplierAssessment(accessToken, referralId)
-    const intervention = await this.interventionsService.getIntervention(accessToken, referral.referral.interventionId)
-    const deliusOfficeLocations: DeliusOfficeLocation[] =
-      await this.deliusOfficeLocationFilter.findOfficesByIntervention(intervention)
-    const { currentAppointment } = new SupplierAssessmentDecorator(supplierAssessment)
-    const hasExistingScheduledAppointment = currentAppointment !== null && !currentAppointment.sessionFeedback.submitted
-
-    await this.scheduleAppointment(
-      req,
-      res,
-      draft,
-      {
-        getReferral: async () => referral,
-        scheduleAppointment: paramsForUpdate =>
-          this.interventionsService
-            .scheduleSupplierAssessmentAppointment(
-              res.locals.user.token.accessToken,
-              supplierAssessment.id,
-              paramsForUpdate
-            )
-            .then(),
-        createPresenter: async (formError, userInputData, serverError) => {
-          let assignedCaseworker: AuthUserDetails | null = null
-          if (currentAppointment?.sessionFeedback?.submitted) {
-            assignedCaseworker = await this.hmppsAuthService.getSPUserByUsername(
-              res.locals.user.token.accessToken,
-              currentAppointment!.sessionFeedback!.submittedBy!.username
-            )
-          }
-          let appointmentSummary: AppointmentSummary | null = null
-          if (currentAppointment) {
-            const deliusOfficeLocation = await this.deliusOfficeLocationFilter.findOfficeByAppointment(
-              currentAppointment
-            )
-            appointmentSummary = new AppointmentSummary(currentAppointment, assignedCaseworker, deliusOfficeLocation)
-          }
-          const overrideBackLinkHref = hasExistingScheduledAppointment
-            ? `/service-provider/referrals/${referralId}/supplier-assessment`
-            : undefined
-          return new ScheduleAppointmentPresenter(
-            'supplierAssessment',
-            referral,
-            currentAppointment,
-            appointmentSummary,
-            deliusOfficeLocations,
-            formError,
-            draft.data,
-            userInputData,
-            serverError,
-            overrideBackLinkHref
-          )
-        },
-        redirectTo: `/service-provider/referrals/${referralId}/supplier-assessment/schedule/${draft.id}/check-answers`,
-      },
-      deliusOfficeLocations
-    )
-  }
-
-  async checkSupplierAssessmentAnswers(req: Request, res: Response): Promise<void> {
-    const referralId = req.params.id
-
-    const referral = await this.interventionsService.getSentReferral(res.locals.user.token.accessToken, referralId)
-    const serviceUser = await this.communityApiService.getServiceUserByCRN(referral.referral.serviceUser.crn)
-
-    const fetchResult = await this.fetchDraftBookingOrRenderMessage(req, res)
-    if (fetchResult.rendered) {
-      return
-    }
-
-    const presenter = new InitialAssessmentCheckAnswersPresenter(fetchResult.draft, referral.id)
-    const view = new ScheduleAppointmentCheckAnswersView(presenter)
-
-    ControllerUtils.renderWithLayout(res, view, serviceUser)
-  }
-
-  async submitSupplierAssessment(req: Request, res: Response): Promise<void> {
-    const referralId = req.params.id
-    const { draftBookingId } = req.params
-    const { accessToken } = res.locals.user.token
-    const supplierAssessment = await this.interventionsService.getSupplierAssessment(accessToken, referralId)
-
-    const { currentAppointment } = new SupplierAssessmentDecorator(supplierAssessment)
-    const hasExistingScheduledAppointment = currentAppointment !== null && !currentAppointment.sessionFeedback.submitted
-
-    await this.submitAppointment(req, res, {
-      scheduleAppointment: paramsForUpdate =>
-        this.interventionsService
-          .scheduleSupplierAssessmentAppointment(
-            res.locals.user.token.accessToken,
-            supplierAssessment.id,
-            paramsForUpdate
-          )
-          .then(),
-      redirectToOnSuccess: `/service-provider/referrals/${referralId}/supplier-assessment/${
-        hasExistingScheduledAppointment ? 'rescheduled-confirmation' : 'scheduled-confirmation'
-      }`,
-      redirectToOnClash: `/service-provider/referrals/${req.params.id}/supplier-assessment/schedule/${draftBookingId}/details?clash=true`,
-    })
-  }
-
-  async showSupplierAssessmentAppointment(req: Request, res: Response): Promise<void> {
-    const referralId = req.params.id
-
-    const [referral, supplierAssessment] = await Promise.all([
-      this.interventionsService.getSentReferral(res.locals.user.token.accessToken, referralId),
-
-      this.interventionsService.getSupplierAssessment(res.locals.user.token.accessToken, referralId),
-    ])
-
-    const appointment = new SupplierAssessmentDecorator(supplierAssessment).currentAppointment
-    if (appointment === null) {
-      throw new Error('Attempting to view supplier assessment without a current appointment')
-    }
-    const deliusOfficeLocation = await this.deliusOfficeLocationFilter.findOfficeByAppointment(appointment)
-
-    const serviceUser = await this.communityApiService.getServiceUserByCRN(referral.referral.serviceUser.crn)
-
-    const presenter = new SupplierAssessmentAppointmentPresenter(
-      referral,
-      appointment,
-      new AppointmentSummary(appointment, null, deliusOfficeLocation),
-      {
-        userType: 'service-provider',
-      }
-    )
-    const view = new SupplierAssessmentAppointmentView(presenter)
-
-    return ControllerUtils.renderWithLayout(res, view, serviceUser)
-  }
-
-  private async scheduleAppointment(
-    req: Request,
-    res: Response,
-    draft: Draft<DraftAppointmentBooking>,
-    appointmentConfig: {
-      getReferral: () => Promise<SentReferral>
-      scheduleAppointment: (paramsForUpdate: AppointmentSchedulingDetails) => Promise<void>
-      createPresenter: (
-        validationError: FormValidationError | null,
-        userInputData: Record<string, unknown> | null,
-        serverError: FormValidationError | null
-      ) => Promise<ScheduleAppointmentPresenter>
-      redirectTo: string
-    },
-    deliusOfficeLocations: DeliusOfficeLocation[]
-  ): Promise<void> {
-    let userInputData: Record<string, unknown> | null = null
-    let formError: FormValidationError | null = null
-    let serverError: FormValidationError | null = null
-
-    if (req.method === 'POST') {
-      const data = await new ScheduleAppointmentForm(req, deliusOfficeLocations).data()
-
-      if (data.error) {
-        res.status(400)
-        formError = data.error
-        userInputData = req.body
-      } else {
-        await this.draftsService.updateDraft(draft.id, data.paramsForUpdate, { userId: res.locals.user.userId })
-        return res.redirect(appointmentConfig.redirectTo)
-      }
-    } else if (req.query.clash === 'true') {
-      serverError = {
-        errors: [
-          {
-            formFields: ['session-input'],
-            errorSummaryLinkedField: 'session-input',
-            message:
-              'The proposed date and time you selected clashes with another appointment. Please select a different date and time.',
-          },
-        ],
-      }
-    }
-
-    const referral = await appointmentConfig.getReferral()
-    const serviceUser = await this.communityApiService.getServiceUserByCRN(referral.referral.serviceUser.crn)
-
-    const presenter = await appointmentConfig.createPresenter(formError, userInputData, serverError)
-    const view = new ScheduleAppointmentView(presenter)
-    return ControllerUtils.renderWithLayout(res, view, serviceUser)
-  }
-
-  private async fetchDraftBookingOrRenderMessage(req: Request, res: Response) {
-    return ControllerUtils.fetchDraftOrRenderMessage<DraftAppointmentBooking>(req, res, this.draftsService, {
-      idParamName: 'draftBookingId',
-      notFoundUserMessage:
-        'Too much time has passed since you started booking this appointment. Your answers have not been saved, and you will need to start again.',
-      typeName: 'booking',
-    })
-  }
-
-  private async submitAppointment(
-    req: Request,
-    res: Response,
-    appointmentConfig: {
-      scheduleAppointment: (paramsForUpdate: AppointmentSchedulingDetails) => Promise<void>
-      redirectToOnSuccess: string
-      redirectToOnClash: string
-    }
-  ): Promise<void> {
-    const fetchResult = await this.fetchDraftBookingOrRenderMessage(req, res)
-    if (fetchResult.rendered) {
-      return
-    }
-    const { draft } = fetchResult
-
-    if (draft.data === null) {
-      throw new Error('Draft data was unexpectedly null when submitting appointment')
-    }
-
-    try {
-      await appointmentConfig.scheduleAppointment(draft.data)
-    } catch (e) {
-      const interventionsServiceError = e as InterventionsServiceError
-      if (interventionsServiceError.status === 409) {
-        res.redirect(appointmentConfig.redirectToOnClash)
-        return
-      }
-
-      throw e
-    }
-
-    await this.draftsService.deleteDraft(draft.id, { userId: res.locals.user.userId })
-    res.redirect(appointmentConfig.redirectToOnSuccess)
   }
 
   async addPostSessionAttendanceFeedback(req: Request, res: Response): Promise<void> {
